Move new files to folders and ignore one test
